+++ src/main/kotlin/com/arkhamusserver/arkhamus/view/dto/ingame/RecipeDto.kt
@@ -6,7 +6,7 @@ data class RecipeDto(
     var recipeId: Int? = null,
     var item: ItemInformationDto? = null,
     var ingredients: List<IngredientDto>? = emptyList(),
-    var numberOfItems: Long = 1,
+    var numberOfItems: Int = 1,
     var timeToCraft: Long = 0L,
     var crafterTypes: List<CrafterType> = emptyList(),
 )
