+++ src/main/kotlin/com/arkhamusserver/arkhamus/view/dto/netty/response/parts/InventoryCell.kt
@@ -4,5 +4,5 @@ import com.arkhamusserver.arkhamus.model.enums.ingame.Item
 
 data class InventoryCell(
     var itemId: Int = Item.PURE_NOTHING.id,
-    var number: Long = 0L
+    var number: Int = 0
 )
